Splitting windows: get things right.
- Set `split-window-preferred-function` for every supported Emacs version.
- Properly inhibit using the `split-window` window parameter in Emacs >= 27.2.

There are two separate cases of vertical window splitting: manually with
`split-window` and friends (specifically `split-window-right`), and
automatically, when `display-buffer` is invoked.

Both are by default impossible if the margins are too wide, but the way to
handle them differs. Manual splitting is controlled by the window parameters
`split-window` (Emacs 25 and up) and `min-margins` (also Emacs 25 and up, but a
bug prevented it from working until Emacs 27.2; see Emacs bug 44483). The use of
these two window parameters should be dependent on the Emacs version (see Github
issue #47).

Automatic splitting by `display-buffer` is controlled by
`split-window-preferred-function`. It is set to
`visual-fill-column-split-window-sensibly`, which should not be dependent on the
Emacs version. Unfortunately, I had mixed them up and made it dependent of the
version anyway.

This should be fixed now.